reload token after refresh

+++ tournaments/tournament.go
@@ -8,7 +8,7 @@ import (
 
 type Tournament struct {
 	ID        bson.ObjectId `json:"id" bson:"_id,omitempty" update:"nonzero"`
-	Owner     int           `json:"owner" bson:"owner" create:"nonzero" update:"nonzero"`
+	Owner     int64         `json:"owner" bson:"owner" create:"nonzero" update:"nonzero"`
 	Name      string        `json:"name" create:"nonzero" update:"nonzero"`
 	Format    int           `json:"format" create:"nonzero" update:"nonzero"`
 	Slots     int           `json:"slots" create:"nonzero" update:"nonzero"`
